+++ feature/login/src/main/java/com/yapp/gallery/login/navigation/LoginNavigatorImpl.kt
@@ -2,7 +2,7 @@ package com.yapp.gallery.login.navigation
 
 import android.content.Context
 import android.content.Intent
-import com.yapp.gallery.login.LoginActivity
+import com.yapp.gallery.login.screen.LoginActivity
 import com.yapp.gallery.navigation.login.LoginNavigator
 import javax.inject.Inject
