+++ feature/login/src/main/java/com/yapp/gallery/login/screen/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.yapp.gallery.login
+package com.yapp.gallery.login.screen
 
 import android.annotation.SuppressLint
 import android.content.res.Configuration
@@ -9,20 +9,19 @@ import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.yapp.gallery.common.theme.*
+import com.yapp.gallery.login.R
 
 @Composable
 fun LoginScreen(
