+++ feature/login/src/main/java/com/yapp/gallery/login/screen/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.yapp.gallery.login
+package com.yapp.gallery.login.screen
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
@@ -9,7 +9,6 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
-import org.apache.commons.lang3.mutable.Mutable
 import javax.inject.Inject
 
 @HiltViewModel
